requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
matplotlib>=3.7.0
numpy>=1.24.0
google-cloud-storage>=2.7.0
asyncio-tools>=0.1.0
python-dotenv>=1.0.0